Address copilot pass 5 on #400
Two findings addressed, one rejected (again).

1. Duplication between lws-cid.js and nostr.js fetchers (line 409).
   Real maintainability concern — every SSRF / redirect / size-cap
   hardening fix would have to land twice. Extracted the shared
   logic to a new module src/auth/cid-doc-fetch.js exporting
   `fetchCidDocument(url, opts)`. Both callers now reduce to a
   one-line delegate. Net: -65 LOC across the two callers, +129
   LOC in the new shared module (with full doc-comment), single
   source of truth for security-critical fetch behavior.

2. Comment about port-round-trip was misleading (line 348). The
   single-user and path-mode branches do preserve the port via
   hostRaw, but the subdomain-mode and base-domain branches drop
   it (matching buildResourceUrl's canonicalization). Comment
   rewritten to be specific about which branches preserve the
   port and why the others don't.

3. Rejected (again): claim that URL.hostname is bracket-free for
   IPv6 (line 365). Verified empirically on Node 24.5.0:

     > new URL('https://[2001:db8::1]:8443/path').hostname
     '[2001:db8::1]'

   This matches WHATWG URL spec §host serializing rule for IPv6
   addresses. The bracket-stripping is correct; without it the
   subdomain-baseDomain comparison would never match for an IPv6
   deployment. Same false claim as pass 4.

Module test counts unchanged: 47 (lws-cid) + 15 (nostr-cid-vm) = 62
both still passing. Full suite: 700/700, no regressions.
