+++ src/auth/lws-cid.js
@@ -35,7 +35,7 @@
 import * as jose from 'jose';
 import { secp256k1 } from '@noble/curves/secp256k1';
 import { sha256 } from '@noble/hashes/sha256';
-import { validateExternalUrl } from '../utils/ssrf.js';
+import { fetchCidDocument } from './cid-doc-fetch.js';
 
 // JWS algorithms we accept. ES256K (RFC8812) is the primary target —
 // secp256k1, the same curve as Nostr — but we support the common JWS
@@ -65,12 +65,10 @@ const PROFILE_CACHE_TTL = 5 * 60 * 1000; // 5 minutes for hits
 const PROFILE_FAILURE_TTL = 60 * 1000;   // 1 minute for misses
 const PROFILE_CACHE_MAX = 1000;          // simple LRU bound
 
-// Manual-redirect cap so a chain can't loop or grind.
-const MAX_REDIRECTS = 5;
-
-// Max profile body size — guards against DoS via giant JSON bodies on
-// untrusted URLs. CID documents are tiny in practice (~1-5 KB).
-const MAX_PROFILE_BYTES = 256 * 1024; // 256 KB
+// Max profile body size — passed to the shared fetcher. CID documents
+// are tiny in practice (~1-5 KB); 256 KB leaves plenty of headroom
+// while bounding any DoS attempt.
+const MAX_PROFILE_BYTES = 256 * 1024;
 
 /** @internal — exposed for tests */
 export function _clearProfileCacheForTests() {
@@ -469,123 +467,16 @@ function setCached(url, entry) {
 }
 
 /**
- * Fetch the CID document with SSRF protection.
+ * Fetch the CID document with SSRF protection. Delegates to the
+ * shared `fetchCidDocument` helper so the per-request defenses (SSRF
+ * validation per hop, manual redirects, same-origin enforcement,
+ * body-size cap) live in one place — see src/auth/cid-doc-fetch.js.
  *
- * docUrl comes from JWT claims (sub, kid) BEFORE the signature is
- * verified, so it's untrusted. We:
- *   1. Validate it through the existing SSRF guard (blocks loopback,
- *      private IPs, http (in production), DNS that resolves to private
- *      addresses).
- *   2. Disable automatic redirects and re-validate every Location to
- *      defeat redirect-based bypasses (mirrors the cors-proxy pattern).
- *      Cross-origin redirects are refused — otherwise a target
- *      attacker-controlled host could serve a substitute CID document
- *      for the WebID's origin.
- *   3. Cap redirects so a chain can't loop.
- *   4. Cap response body size so a giant payload can't OOM us.
- *   5. Always send a fresh Accept and a small read-side timeout.
+ * The cache wrapper around this function (above) is LWS-CID-specific
+ * and stays here.
  */
 async function fetchProfileNoCache(docUrl) {
-  const originalOrigin = new URL(docUrl).origin;
-  let currentUrl = docUrl;
-
-  // Hop 0 is the original request; up to MAX_REDIRECTS subsequent
-  // redirects are followed, after which we throw.
-  for (let hop = 0; hop <= MAX_REDIRECTS; hop++) {
-    const isLastAllowedHop = hop === MAX_REDIRECTS;
-    const validation = await validateExternalUrl(currentUrl, {
-      // Allow http on dev only — production deploys should always be https.
-      requireHttps: process.env.NODE_ENV === 'production',
-      blockPrivateIPs: true,
-      resolveDNS: true,
-    });
-    if (!validation.valid) {
-      throw new Error(`SSRF protection: ${validation.error}`);
-    }
-
-    const controller = new AbortController();
-    const timer = setTimeout(() => controller.abort(), 5000);
-    let res;
-    try {
-      res = await fetch(currentUrl, {
-        // Prefer JSON-LD but accept plain JSON too — some WebID hosts
-        // serve `application/json` for `card.jsonld`. The body is JSON
-        // either way; we don't perform JSON-LD-specific processing here.
-        headers: { Accept: 'application/ld+json, application/json;q=0.9' },
-        redirect: 'manual',
-        signal: controller.signal,
-      });
-    } finally {
-      clearTimeout(timer);
-    }
-
-    // Manual redirect handling — re-validate every Location and require
-    // same-origin so a redirect can't substitute an attacker-controlled
-    // CID document for the WebID's origin.
-    if (res.status >= 300 && res.status < 400) {
-      if (isLastAllowedHop) {
-        throw new Error(`too many redirects (>${MAX_REDIRECTS})`);
-      }
-      const loc = res.headers.get('location');
-      if (!loc) throw new Error(`redirect ${res.status} without Location`);
-      const nextUrl = new URL(loc, currentUrl).toString();
-      const nextOrigin = new URL(nextUrl).origin;
-      if (nextOrigin !== originalOrigin) {
-        throw new Error(
-          `cross-origin redirect refused: ${originalOrigin} → ${nextOrigin}`,
-        );
-      }
-      currentUrl = nextUrl;
-      continue;
-    }
-
-    if (!res.ok) {
-      throw new Error(`HTTP ${res.status}`);
-    }
-
-    // Size guard. Two layers: trust Content-Length when present, then
-    // also enforce as we read so a streaming response can't lie.
-    const declared = Number(res.headers.get('content-length'));
-    if (Number.isFinite(declared) && declared > MAX_PROFILE_BYTES) {
-      throw new Error(
-        `CID document too large (Content-Length=${declared} > ${MAX_PROFILE_BYTES})`,
-      );
-    }
-    const text = await readBodyWithCap(res, MAX_PROFILE_BYTES);
-    return JSON.parse(text);
-  }
-  // Loop exited without returning or redirecting — defensive fallback.
-  throw new Error('profile fetch loop exited unexpectedly');
-}
-
-/**
- * Read response body with a hard byte cap. Aborts the stream as soon as
- * the cap is exceeded so we don't buffer the entire untrusted payload.
- */
-async function readBodyWithCap(res, maxBytes) {
-  const reader = res.body?.getReader?.();
-  if (!reader) {
-    // No streaming reader (older runtimes / mocked responses) — fall
-    // back to .text() but enforce the cap after the fact.
-    const text = await res.text();
-    if (Buffer.byteLength(text, 'utf8') > maxBytes) {
-      throw new Error(`CID document too large (>${maxBytes} bytes)`);
-    }
-    return text;
-  }
-  const chunks = [];
-  let total = 0;
-  for (;;) {
-    const { value, done } = await reader.read();
-    if (done) break;
-    total += value.byteLength;
-    if (total > maxBytes) {
-      try { await reader.cancel(); } catch { /* noop */ }
-      throw new Error(`CID document too large (>${maxBytes} bytes)`);
-    }
-    chunks.push(value);
-  }
-  return Buffer.concat(chunks.map((c) => Buffer.from(c.buffer, c.byteOffset, c.byteLength))).toString('utf8');
+  return fetchCidDocument(docUrl, { maxBytes: MAX_PROFILE_BYTES });
 }
 
 function findVerificationMethod(profile, kid, baseUrl) {
